bug 8128; Fix common error/warn dialogs to deal with modal parent
Unmanaged dialogs are often called for errors, warnings, info etc.
On Windows, If called from a modal parent window, they can be a bit
confusing, since the parent window appears to still be active.  This
fix makes the parent window non-modal for the duration of the dialog,
and then returns it to its original state.