bug 9855; fix db undo/redo operation to delay signal emission
Certain multiple commit transactions can leave the db in an inconsistent
state in between commits.  If signals are emitted at each commit, GUI
elements can see the inconsistent state and report errors.  This fix
delays the signal emission until all the commits are complete, presuming
that the db is consistent before and after the complete transaction.